requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis Cache
redis==5.0.1

# Google Trends
pytrends==4.9.2

# Environment & Configuration
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1

# Logging
loguru==0.7.2

# HTTP & Rate Limiting
Flask-Limiter==3.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
requests-mock==1.11.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Production Server
gunicorn==21.2.0
